앱바 리팩토링
