+++ src/pageImpl/mainImpl/__components__/MainAppBar.tsx
@@ -10,7 +10,7 @@ export const MainAppBar = () => {
   const router = useRouter();
 
   return (
-    <AppBar leftImage={<SvgTimetableOn height={30} width={30} />}>
+    <AppBar left={<SvgTimetableOn height={30} width={30} />}>
       <AppBarContent>
         <Title>강의평</Title>
         <SvgSearchOff data-testid="main-search-icon" height={30} width={30} onClick={() => router.push('/search')} />
